% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ReadGMT}
\alias{ReadGMT}
\title{Parse a GMT file into a data frame}
\usage{
ReadGMT(gmtpath, limit = c(1, Inf))
}
\arguments{
\item{gmtpath}{Path to a GMT-format file.}

\item{limit}{A two-length numeric vector giving the minimum and maximum
gene-set size to retain (inclusive on both ends).}
}
\value{
A data frame with columns \code{Gene}, \code{PathwayID}, and
  \code{PathwayName}.
}
\description{
Parse a GMT file into a data frame
}
