% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsGetter.R
\docType{methods}
\name{retrieve_gs}
\alias{retrieve_gs}
\title{Update genesets from source database}
\usage{
retrieve_gs(
  type = c("KEGG", "REACTOME", "CORUM", "GO", "MSIGDB"),
  organism = "hsa",
  msigdb.path = NULL,
  release = NULL
)
}
\arguments{
\item{type}{A vector of databases, such as KEGG, REACTOME, CORUM, GO, MSIGDB}

\item{organism}{'hsa' or 'mmu'.}

\item{msigdb.path}{Path to msigdb sqlite file. Must be provided if `MSIGDB` is specified in `type`}
}
\value{
save data to local library.
}
\description{
Update genesets from source database
}
\author{
Wubing Zhang
}
