% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransGeneID.R
\docType{methods}
\name{getGeneAnn}
\alias{getGeneAnn}
\title{Retrieve gene annotations from the NCBI, HNSC, and Uniprot databases.}
\usage{
getGeneAnn(org = "hsa", update = FALSE, release = NULL)
}
\arguments{
\item{org}{Character, hsa (default), bta, cfa, mmu, ptr, rno, ssc are optional.}

\item{update}{Boolean, indicating whether download current annotation.}
}
\value{
A data frame.
}
\description{
Retrieve gene annotations from the NCBI, HNSC, and Uniprot databases.
}
\examples{
\dontrun{
  ann = getGeneAnn("hsa")
  head(ann)
}

}
\author{
Wubing Zhang
}
