% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsGetter.R
\docType{methods}
\name{gsGetter}
\alias{gsGetter}
\title{Extract pathway annotation from GMT file.}
\usage{
gsGetter(
  gmtpath = NULL,
  type = "All",
  limit = c(0, Inf),
  organism = "hsa",
  update = FALSE,
  msigdb.path = NULL
)
}
\arguments{
\item{gmtpath}{The path to customized gmt file.}

\item{type}{Molecular signatures for testing, available datasets include
Pathway (KEGG, REACTOME, C2_CP_PID, C2_CP_BIOCARTA, C2_CP_WIKIPATHWAYS, C2_CP_KEGG_MEDICUS), GO (GOBP, GOCC, GOMF),
MSIGDB (C1, C2 (C2_CP (C2_CP_PID, C2_CP_BIOCARTA, C2_CP_WIKIPATHWAYS, C2_CP_KEGG_MEDICUS), C2_CGP),
C3 (C3_MIR_MIRDB, C3_TFT_GTRD), C4 (C4_CGN, C4_CM, c4_3CA), C5 (C5_GO_BP, C5_GO_CC, C5_GO_MF, C5_HPO), C6, C7 (C7_IMMUNESIGDB, C7_VAX), C8, H)
and Complex (CORUM). Any combination of them are also accessible
(e.g. 'GOBP+GOMF+KEGG+REACTOME').}

\item{limit}{A two-length vector, specifying the minimal and
maximal size of gene sets to load.}

\item{organism}{'hsa' or 'mmu'.}

\item{update}{Boolean, indicating whether update the gene sets from source database.}

\item{msigdb.path}{Path to msigdb sqlite file. Must be provided if `update` = `TRUE` and "MSIGDB" is specified in `type`}
}
\value{
A three-column data frame.
}
\description{
Extract pathway annotation from GMT file.
}
\examples{
gene2path = gsGetter(type = "REACTOME+KEGG")
head(gene2path)

}
\author{
Wubing Zhang
}
